Syn data speed master merge (#25)
Added a baseline called IPBatching that chooses predicates using the ticketing system, but does not use a queue and completely finishes the chosen IP pair before moving on to the next one.

Added a toggle to generate a graph that shows how the number of tickets for each predicate changes every time a task is assigned.

Changed the "switching" functionality so  that you can input how much you want to change each predicate's selectivity and ambiguity by, and specify at point in the simulation you want that to happen. Also added the option to have the sel/amb change in the form of a SIN curve.

Added a toggle to generate a graphs that shows each active predicate's selectivities over the course of one simulation. This is currently being used to make sure the predicates' selectivities are indeed changing in the manner expected (ex. SIN curve).
Changed code for synthetic data to be more model-based, instead of calling .save() on a bunch of objects within the code

**Note: The switching and IPBatching functionality does not yet work for time simulations.